feat: implement user authentication with registration, login, and session management

+++ src/controllers/notesController.js
@@ -1,5 +1,5 @@
 import createHttpError from 'http-errors';
-import { Note } from '../models/note.js';
+import Note from '../models/note.js';
 
 export const getAllNotes = async (req, res) => {
   const page = Number(req.query.page) || 1;
@@ -7,24 +7,25 @@ export const getAllNotes = async (req, res) => {
   const { tag, search } = req.query;
   const skip = (page - 1) * perPage;
 
-  const notesQuery = Note.find();
+  const filter = {};
 
   if (tag) {
-    notesQuery.where('tag').equals(tag);
+    filter.tag = tag;
   }
-
   if (search && search.trim() !== '') {
-    notesQuery.where({
-      $or: [
-        { title: { $regex: search, $options: 'i' } },
-        { content: { $regex: search, $options: 'i' } },
-      ],
-    });
+    filter.$text = { $search: search };
   }
 
-  const [totalNotes, notes] = await Promise.all([
-    notesQuery.clone().countDocuments(),
-    notesQuery.skip(skip).limit(perPage).exec(),
+  // if (search && search.trim() !== '') {
+  //   filter.$or = [
+  //     { title: { $regex: search, $options: 'i' } },
+  //     { content: { $regex: search, $options: 'i' } },
+  //   ];
+  // }
+
+  const [notes, totalNotes] = await Promise.all([
+    Note.find({ userId: req.user._id, ...filter }).skip(skip).limit(perPage).exec(),
+    Note.countDocuments({ userId: req.user._id, ...filter }),
   ]);
 
   const totalPages = Math.ceil(totalNotes / perPage);
@@ -40,22 +41,21 @@ export const getAllNotes = async (req, res) => {
 
 export const getNoteById = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findById(noteId);
+  const note = await Note.findOne({ _id: noteId, userId: req.user._id });
   if (!note) {
-    next(createHttpError(404, 'Note not found'));
-    return;
+    return next(createHttpError(404, 'Note not found'));
   }
   res.status(200).json(note);
 };
 
 export const createNote = async (req, res) => {
-  const note = await Note.create(req.body);
+  const note = await Note.create({ ...req.body, userId: req.user._id });
   res.status(201).json(note);
 };
 
 export const deleteNote = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findOneAndDelete({ _id: noteId });
+  const note = await Note.findOneAndDelete({ _id: noteId, userId: req.user._id });
   if (!note) {
     next(createHttpError(404, 'Note not found'));
     return;
@@ -65,8 +65,8 @@ export const deleteNote = async (req, res, next) => {
 
 export const updateNote = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findOneAndUpdate({ _id: noteId }, req.body, {
-    returnDocument: 'after',
+  const note = await Note.findOneAndUpdate({ _id: noteId, userId: req.user._id }, req.body, {
+    new: true,
   });
   if (!note) {
     next(createHttpError(404, 'Note not found'));
